Helper moved to utils
