+++ packages/core/src/components/badge/badge.tsx
@@ -2,7 +2,7 @@
 import { Component, Host, h, Prop } from '@stencil/core';
 
 import type { Color } from '../../types';
-import { createColorClasses } from '../../utils';
+import { createColorClasses } from '../../utils/utils';
 
 @Component({
   tag: 'stories-badge',
