+++ packages/core/src/custom-elements.json
@@ -1,5 +1,5 @@
 {
-  "timestamp": "2022-04-15T18:27:38",
+  "timestamp": "2022-04-16T07:46:32",
   "compiler": {
     "name": "@stencil/core",
     "version": "2.14.2",
