+++ packages/core/src/index.html
@@ -806,7 +806,7 @@
       <h2>Grid (stories-grid)</h2>
 
       <section>
-        <h3>One Row</h3>
+        <h3>Basic</h3>
 
         <stories-grid>
           <stories-row>
